+++ CHANGELOG.md
@@ -39,6 +39,7 @@ FEATURES:
   * Supports the new fields (`hmac_accessor`, `jsonx` format etc.) for File and SysLog Audit Backends.
   * The `AWSGenerateDynamicCredentialsWithSecurityTokenAsync` method now supports the `timeToLive` parameter.
   * The `Consul` backend now supports the listing functionality to roles `ConsulReadRoleListAsync`. (https://github.com/hashicorp/vault/issues/2065)
+  * The `Transit` backend now supports the new Apis for `List of keys`, `Random`, `Hash`, `Digest`, `Sign`, `Verify` etc.
   * All the secret backends now support wrapping of the secret into a cubbyhole token. Wrapping support added for:
     * AWS Secret Backend
     * Cassandra Secret Backend